start working on step automation
step is the relation that moves a head position reduction to
any place in the AST where we might need it. To do this lifting
we use a context and therefore from a head_step proof we get a step
proof only if we rewrite a lot of the AST in this form.

Obviously this is very boring so automation is valuable here.