add error frame rule
I had to modify the error definition but I proved the frame rule.
The hard part was getting to understand _how_ to state that during
the reduction the program was not allowed to touch the frame.

This intuition comes from the idea of the _footprint_ of the
program, all resources that are used by the program, and having
read that programs are characterized by this _footprint theorem_.