unify wp and ewp
Robbert suggested unifying wp and ewp by having a match on the final
value. This helps with the amount of lemmas that I had to prove as
now there is only one predicate and also evaluating in context holes
has been taken care of.

This predicate also relies on frame baking as the previous two and we
might rework this when we hit some more problems.