update heap definition
our current heap definition is a partial map that does not preserve
freed cells. This is gonna be problematic later on when we want to
prove the heap manipulation proof rules.

This changes the underlying value in the map from a value to a
disjoint union of a value or a Reserved cell.

Reserved is a value that only the EFree expression can put in the
heap and it is used to distinguish cells that were previously allocated
and freed from cells that were never allocated.

The head_step rule for EAlloc then changes to allow not only
allocating unused cells but the Reserved too. The head_step rule
for EFree, EStore, ELoad instead change the lookup value to be
Some (Value w) because we have to restric us from loading previously
deallocated cells.

The Alloc_head_step rule could be split into two, one for each result of
lookup that we allow, None and Some (Reserved) but this can be
unified using another predicate.

m !! l = Some e -> e = Reserved gives us both the allowed cases;
if location l has value e then it is a Reserve, meaning that it
was previously freed.