refactor proofs of post rules
many of the proofs done for proving the strongest post
entailments were done at the _unfolding level_ in one
way or the other.

Since Jules proposed to add the pure_step and post_pure_step
rules we saw that many of the _derived rules_ could be proved
with more primitive rules of post and entailment such as post_mono
and iEntails_trans.

This path does exactly that and proves the usefulness of the
post_pure_step rule.