add both versions of the hoare sequence rule
Robbert wants me to prove the more general rule for the sequence where
we are allowed to not pick a value to put in the sequence as we will
discard it.

This rule should not be provable with the current hoare_ctxS rule as
we will still have a proof obligation in the form of an existential variable.