adds the hoare context rule taken from Iris
This rule is different from the previous hoare context
rule because the quantification of v is at the triple
level instead of the lemma level.

This means that we can still use the predicate P' to make
an assertion on the possible result values of e while at the
same time we are not required to give it out at the lemma level.