update hoare rules for nicer experience when using them
the hoare rules arguments must be given in order and this
means that the rules are nicer to use if the arguments that
can be inferred automatically are not the first.

This means that we can be specific with the first arguments
or not and the rest will follow along.