add a hoare context triple that mimics the SL one
In separation logic the hoare triple for the context is as follows

{{ P }} e {{ x, R x }}  forall v, {{ R v }} K[v] {{ x, Q x}}
------------------------------------------------------------
	{{ P }} e {{ x, Q x}}

But this rule makes use of the fact that {{ False }} e {{x, Q x}}
is always valid.

In IL this is not true, the trivial triple is  [ P ] e [x, False ]
which gave me the idea of moving the value v to the result assertion.

This is actually in line with the previous IL triples that I saw where
the distinction on the good/bad values are put in the result
assertion.

Say K ≝ let x = ⬜ in if x ≥ 1 then x else error. Then the traditional
incorrectness triples are written as.

[v ∈ N] K[v] [r, v ≥ 1]
[v ∈ N] K[v] [ERR: v = 0]

So the current hoare_ctxS_iris_forall' rule encodes this pattern
as we can always split the bad values from the good value in the
result assertion.