rename hoare_pure to hoare_intro
this rule mimicks the iIntro rule from Iris. Pure assertions in the
precondition of a correctness triple can be pulled in the proof
context and used when unfolding a program execution but in our logic
it's not the pure assertions in the presumption but the pure assertion
in the result that can be pulled.

This also is a path I've been thinking on recently; an incorrectness
triple {{ ⌜ Φ ⌝ ∗ P }} e {{ Q }} cannot give you the Φ?