prove context rule for error triples in Iris style
The new rules are a generalisation of the normal hoare
context rule and to prove it we show how to recover
the first context rule from the general.

Both rules admit the existence of values v in the lemma
premises but they describe it differently.

If in the first triple the value was explicit when using
⌜ r = v ⌝ or ⌜ Φ v ⌝ in the triple result for the general
rule we don't need to be explicity immediately.

There is a problem though as now the result assertion for
is Q : val → val → iProp instead of being Q :  val → iProp
and this might confuse people. Where do we get the new v
in the current triple if it does not have it?

You can always introduce a dummy ⌜ Φ v ⌝ in the result assertion
or just ignore v by wrapping in another lamba the original term
λ _ : Q.